+++ tests/fixtures.py
@@ -3,6 +3,7 @@
 #  (c) 2022-2023. A. Shavykin <0.delameter@gmail.com>
 #  Licensed under GNU Lesser General Public License v3.0
 # -----------------------------------------------------------------------------
+
 import pytest
 
 from pytermor import ExtendedEnum, RendererManager