+++ tests/test_style.py
@@ -16,7 +16,12 @@ from pytermor import (
     NOOP_STYLE,
     Color16,
     DEFAULT_COLOR,
-    NOOP_COLOR, Color256, ColorRGB, IColor, make_style, )
+    NOOP_COLOR,
+    Color256,
+    ColorRGB,
+    IColor,
+    make_style,
+)
 from pytermor.exception import LogicError, ArgTypeError
 
 from tests import format_test_params
@@ -26,20 +31,20 @@ class TestStyle:
     @pytest.mark.parametrize(
         "expected, arg",
         [
-            (NOOP_STYLE,            None),
-            (Style(fg='red'),       'red'),
-            (Style(fg='green'),     cv.GREEN),
-            (Style(fg='navy_blue'), cv.NAVY_BLUE),
-            (Style(fg=0x0eb0b0),    0x0eb0b0),
-            (Style(fg=0xaaaaaa),    "#aaa"),
-            (Style(fg=0xa0aaca),    "#a0aaca"),
+            (NOOP_STYLE, None),
+            (Style(fg="red"), "red"),
+            (Style(fg="green"), cv.GREEN),
+            (Style(fg="navy_blue"), cv.NAVY_BLUE),
+            (Style(fg=0x0EB0B0), 0x0EB0B0),
+            (Style(fg=0xAAAAAA), "#aaa"),
+            (Style(fg=0xA0AACA), "#a0aaca"),
             (
-                Style(bg='yellow', bold=True),
-                Style(bg='yellow', bold=True),
-             ),
+                Style(bg="yellow", bold=True),
+                Style(bg="yellow", bold=True),
+            ),
             pytest.param(None, [], marks=pytest.mark.xfail(raises=ArgTypeError)),
         ],
-        ids=format_test_params
+        ids=format_test_params,
     )
     def test_make_style(self, expected: Style, arg: typing.Any):
         assert make_style(arg) == expected
@@ -52,16 +57,16 @@ class TestStyle:
         "expected_result, color_arg",
         [
             (0x800000, cv.RED),
-            (0xaf8700, cv.DARK_GOLDENROD),
+            (0xAF8700, cv.DARK_GOLDENROD),
             (0x800000, Color16(0x800000, IntCode.RED, IntCode.BG_RED)),
-            (0xffaacc, Color256(0xffaacc, 263)),
-            (0xff00cc, ColorRGB(0xff00cc)),
-            (0x00aacc, 0x00aacc),
-            (0x9a9a9a, "#9a9a9a"),
+            (0xFFAACC, Color256(0xFFAACC, 263)),
+            (0xFF00CC, ColorRGB(0xFF00CC)),
+            (0x00AACC, 0x00AACC),
+            (0x9A9A9A, "#9a9a9a"),
             (0x666666, "#666"),
             (0x000000, "black"),
-            (0x00005f, "navy_blue"),
-            (0x0052cc, "jira_blue"),
+            (0x00005F, "navy_blue"),
+            (0x0052CC, "jira_blue"),
             pytest.param(None, [], marks=pytest.mark.xfail(raises=ArgTypeError)),
             pytest.param(None, None, marks=pytest.mark.xfail(raises=LogicError)),
         ],
@@ -76,10 +81,10 @@ class TestStyle:
         "st",
         [
             Style(),
-            Style(fg='red'),
-            Style(bg='green'),
-            Style(fg='blue', bg='black'),
-        ]
+            Style(fg="red"),
+            Style(bg="green"),
+            Style(fg="blue", bg="black"),
+        ],
     )
     def test_flip(self, st: Style):
         st_flipped = st.clone().flip()
@@ -89,28 +94,28 @@ class TestStyle:
     @pytest.mark.parametrize(
         "expected_fg_brightness, bg",
         [
-            (.10, 0xff0000),
-            (.10, 0xffff00),
-            (.10, 0x00ffff),
-            (.10, 0x0000ff),
-            (.10, 0xffffff),
-            (.10, 0x800000),
-            (.10, 0x008000),
-            (.10, 0x000080),  # @TODO fix this
-            (.10, 0x808000),
-            (.10, 0x008080),
-            (.10, 0x808080),
-            (.80, 0x400000),
-            (.80, 0x004000),
-            (.80, 0x000040),
-            (.80, 0x404040),
-            (.80, 0x000000),
-        ]
+            (0.10, 0xFF0000),
+            (0.10, 0xFFFF00),
+            (0.10, 0x00FFFF),
+            (0.10, 0x0000FF),
+            (0.10, 0xFFFFFF),
+            (0.10, 0x800000),
+            (0.10, 0x008000),
+            (0.10, 0x000080),  # @TODO fix this
+            (0.10, 0x808000),
+            (0.10, 0x008080),
+            (0.10, 0x808080),
+            (0.80, 0x400000),
+            (0.80, 0x004000),
+            (0.80, 0x000040),
+            (0.80, 0x404040),
+            (0.80, 0x000000),
+        ],
     )
-    def test_autopick_fg(self, expected_fg_brightness: float, bg: int|None):
+    def test_autopick_fg(self, expected_fg_brightness: float, bg: int | None):
         st = Style(bg=bg).autopick_fg()
         fg_brightness = st.fg.to_hsv().value
-        assert isclose(fg_brightness, expected_fg_brightness, abs_tol=.10)  # 10% margin
+        assert isclose(fg_brightness, expected_fg_brightness, abs_tol=0.10)  # 10% margin
 
     def test_autopick_fg_doesnt_change_without_bg(self):
         assert Style(fg=0x800080).autopick_fg().fg.hex_value == 0x800080
@@ -124,6 +129,10 @@ class TestStyle:
             (Style(bg="red"), Style(bg="red")),
             (Style(fg="red", bg="black"), Style(fg="red", bg="black")),
             (Style(fg="red", bold=True), Style(fg="red", bold=True)),
+            (
+                Style(underline_color="red", bold=True),
+                Style(underline_color="red", bold=True),
+            ),
             (Style(underlined=True, italic=True), Style(underlined=True, italic=True)),
         ],
         ids=format_test_params,
@@ -140,6 +149,7 @@ class TestStyle:
             (Style(bg="red"), Style(fg="red")),
             (Style(fg="red", bg="black"), Style(fg="red", bg="yellow")),
             (Style(fg="red", bold=True), Style(fg="red", bold=False)),
+            (Style(underline_color="red"), Style(underline_color="dark-red")),
             (Style(underlined=True, italic=True), Style(underlined=False, italic=True)),
         ],
         ids=format_test_params,
@@ -168,6 +178,7 @@ class TestStyle:
         [
             ("fg", lambda st: setattr(st, "fg", "blue")),
             ("bg", lambda st: setattr(st, "bg", "red")),
+            ("underline_color", lambda st: setattr(st, "underline_color", "yellow")),
             ("bold", lambda st: setattr(st, "bold", True)),
             ("flip", lambda st: st.flip()),
             ("autopick_fg", lambda st: st.autopick_fg()),
@@ -176,10 +187,13 @@ class TestStyle:
         ],
         ids=format_test_params,
     )
-    @pytest.mark.parametrize("instantiate_fn", [
-        lambda: Style(frozen=True),
-        lambda: NOOP_STYLE.__class__(),
-    ])
+    @pytest.mark.parametrize(
+        "instantiate_fn",
+        [
+            lambda: Style(frozen=True),
+            lambda: NOOP_STYLE.__class__(),
+        ],
+    )
     def test_frozen_style_immutability(
         self,
         label: str,
@@ -195,24 +209,39 @@ class TestStyle:
             ("<Style[]>", Style()),
             ("<*Style[]>", Style(frozen=True)),
             ("<*_NoOpStyle[]>", NOOP_STYLE),
-            ("<Style[red]>", Style(fg='red')),
-            ("<Style[|red]>", Style(bg='red')),
+            ("<Style[red]>", Style(fg="red")),
+            ("<Style[|red]>", Style(bg="red")),
             ("<Style[x88]>", Style(fg=cv.DARK_RED)),
             ("<Style[|x88]>", Style(bg=cv.DARK_RED)),
             ("<Style[#400000]>", Style(fg=0x400000)),
-            ("<Style[|#ffffff]>", Style(bg=0xffffff)),
-            ("<Style[#ff00ff|#008000]>", Style(fg=0xff00ff, bg=0x008000)),
+            ("<Style[|#ffffff]>", Style(bg=0xFFFFFF)),
+            ("<Style[#ff00ff|#008000]>", Style(fg=0xFF00FF, bg=0x008000)),
             ("<Style[#0052cc]>", Style(fg="jira blue")),
             ("<Style[|#ff9966]>", Style(bg="atomic tangerine")),
             ("<Style[#b3f5ff|#824b35]>", Style(fg="arctic chill", bg="green tea")),
+            ("<Style[U:#e7c899]>", Style(underline_color="icathian yellow")),
+            (
+                "<Style[red|blue U:yellow]>",
+                Style(fg="red", bg="blue", underline_color="yellow"),
+            ),
+            ("<Style[red U:yellow]>", Style(fg="red", underline_color="yellow")),
+            ("<Style[|blue U:yellow]>", Style(bg="blue", underline_color="yellow")),
             ("<Style[DEF]>", Style(fg=DEFAULT_COLOR)),
             ("<Style[|DEF]>", Style(bg=DEFAULT_COLOR)),
-            ("<Style[+BOLD +DIM +ITAL +UNDE]>", Style(bold=True, dim=True, italic=True, underlined=True)),
-            ("<Style[+CROS +DOUB +OVER]>", Style(overlined=True, crosslined=True, double_underlined=True)),
+            ("<Style[U:DEF]>", Style(underline_color=DEFAULT_COLOR)),
+            (
+                "<Style[+BOLD +DIM +ITAL +UNDE]>",
+                Style(bold=True, dim=True, italic=True, underlined=True),
+            ),
+            (
+                "<Style[+CROS +DOUB +OVER]>",
+                Style(overlined=True, crosslined=True, double_underlined=True),
+            ),
             ("<Style[+BLIN +INVE]>", Style(inversed=True, blink=True)),
             ("<Style[+DIM -BOLD]>", Style(bold=False, dim=True)),
             ("<Style[red +BOLD]>", Style(fg="red", bold=True)),
             ("<Style[|red -BOLD]>", Style(bg="red", bold=False)),
+            ("<Style[U:yellow -BOLD]>", Style(underline_color="yellow", bold=False)),
         ],
     )
     def test_style_repr(self, expected: str, style: Style):
@@ -221,15 +250,24 @@ class TestStyle:
     @pytest.mark.parametrize(
         "expected, style, verbose",
         [
-            ("c31(#800000? red)", Style(fg='red'), True),
-            ("|c31(#800000? red)", Style(bg='red'), True),
-            ("c30(#000000? black)|c31(#800000? red)", Style(fg=cv.BLACK, bg=cv.RED), True),
+            ("c31(#800000? red)", Style(fg="red"), True),
+            ("|c31(#800000? red)", Style(bg="red"), True),
+            (
+                "c30(#000000? black)|c31(#800000? red)",
+                Style(fg=cv.BLACK, bg=cv.RED),
+                True,
+            ),
+            ("U:c93(#ffff00? hi-yellow)", Style(underline_color=cv.HI_YELLOW), True),
             ("x88(#870000 dark-red)", Style(fg=cv.DARK_RED), True),
             ("|x88(#870000 dark-red)", Style(bg=cv.DARK_RED), True),
-            ("x237(#3a3a3a gray-23)|x88(#870000 dark-red)", Style(fg=cv.GRAY_23, bg=cv.DARK_RED), True),
+            (
+                "x237(#3a3a3a gray-23)|x88(#870000 dark-red)",
+                Style(fg=cv.GRAY_23, bg=cv.DARK_RED),
+                True,
+            ),
             ("#400000", Style(fg=0x400000), True),
-            ("|#ffffff", Style(bg=0xffffff), True),
-            ("#ff00ff|#008000", Style(fg=0xff00ff, bg=0x008000), True),
+            ("|#ffffff", Style(bg=0xFFFFFF), True),
+            ("#ff00ff|#008000", Style(fg=0xFF00FF, bg=0x008000), True),
             ("#0052cc(pacific-bridge)", Style(fg="jira blue"), True),
             ("|#ff9966(atomic-tangerine)", Style(bg="atomic tangerine"), True),
             (
@@ -242,7 +280,12 @@ class TestStyle:
                 Style(fg="arctic chill", bg="green tea", double_underlined=True),
                 False,
             ),
-        ]
+            (
+                "#400000|#ffffff U:#808000",
+                Style(fg=0x400000, bg=0xFFFFFF, underline_color=0x808000),
+                True,
+            ),
+        ],
     )
     def test_style_verbose_repr(self, expected: str, style: Style, verbose: bool):
         assert style.repr_attrs(verbose) == expected
@@ -258,8 +301,12 @@ class TestStyleMerging:
             (Style(fg="yellow"), Style(fg="yellow"), Style(fg="red")),
             (Style(bg="green"), Style(bg="green"), Style(bg=None)),
             (Style(bg=DEFAULT_COLOR), Style(bg=DEFAULT_COLOR), Style(bg=None)),
-            (Style(bg='red', fg='blue'), Style(fg='blue'), Style(bg='red')),
-            (Style(bg='blue', fg='red'), Style(fg='red'), Style(bg='blue')),
+            (Style(bg="red", fg="blue"), Style(fg="blue"), Style(bg="red")),
+            (
+                Style(underlined=False, underline_color="yellow"),
+                Style(underlined=None, underline_color="yellow"),
+                Style(underlined=False, underline_color="green"),
+            ),
             (Style(bold=True), Style(bold=True), Style()),
             (Style(bold=False), Style(bold=False), Style()),
             (Style(bold=True), Style(bold=True), Style(bold=False)),
@@ -271,6 +318,60 @@ class TestStyleMerging:
     )
     def test_style_merging_fallback(self, expected: Style, base: Style, fallback: Style):
         assert base.merge_fallback(fallback) == expected
+
+    @pytest.mark.parametrize(
+        "expected, base, overwrite",
+        [
+            (Style(), Style(), NOOP_STYLE),
+            (Style(fg="red"), Style(fg="red"), NOOP_STYLE),
+            (Style(fg="red"), Style(), Style(fg="red")),
+            (Style(fg="red"), Style(fg="yellow"), Style(fg="red")),
+            (Style(bg="green"), Style(bg="green"), Style(bg=None)),
+            (Style(bg=DEFAULT_COLOR), Style(bg=DEFAULT_COLOR), Style(bg=None)),
+            (Style(bg="red", fg="blue"), Style(fg="blue"), Style(bg="red")),
+            (
+                Style(underlined=False, underline_color="green"),
+                Style(underlined=None, underline_color="yellow"),
+                Style(underlined=False, underline_color="green"),
+            ),
+            (Style(bold=True), Style(bold=True), Style()),
+            (Style(bold=False), Style(bold=False), Style()),
+            (Style(bold=False), Style(bold=True), Style(bold=False)),
+            (Style(bold=True), Style(bold=False), Style(bold=True)),
+            (Style(bold=False), Style(), Style(bold=False)),
+            (Style(bold=True), Style(), Style(bold=True)),
+        ],
+        ids=format_test_params,
+    )
+    def test_style_merging_overwrite(self, expected: Style, base: Style, overwrite: Style):
+        assert base.merge_overwrite(overwrite) == expected
+
+    @pytest.mark.parametrize(
+        "expected, base, replace",
+        [
+            (Style(), Style(), NOOP_STYLE),
+            (Style(), Style(fg="red"), NOOP_STYLE),
+            (Style(fg="red"), Style(), Style(fg="red")),
+            (Style(fg="red"), Style(fg="yellow"), Style(fg="red")),
+            (Style(), Style(bg="green"), Style(bg=None)),
+            (Style(), Style(bg=DEFAULT_COLOR), Style(bg=None)),
+            (Style(bg="red"), Style(fg="blue"), Style(bg="red")),
+            (
+                Style(underlined=False, underline_color="green"),
+                Style(underlined=None, underline_color="yellow"),
+                Style(underlined=False, underline_color="green"),
+            ),
+            (Style(), Style(bold=True), Style()),
+            (Style(), Style(bold=False), Style()),
+            (Style(bold=False), Style(bold=True), Style(bold=False)),
+            (Style(bold=True), Style(bold=False), Style(bold=True)),
+            (Style(bold=False), Style(), Style(bold=False)),
+            (Style(bold=True), Style(), Style(bold=True)),
+        ],
+        ids=format_test_params,
+    )
+    def test_style_merging_replace(self, expected: Style, base: Style, replace: Style):
+        assert base.merge_replace(replace) == expected
 
     def test_noop_style_immutability(self):
         with pytest.raises(LogicError, match="is immutable"):